+++ SDLGame2/SDLGame2/TextureManager.cpp
@@ -43,10 +43,10 @@ SDL_Texture* TextureManager::LoadTexture(std::string _textureID)
 	return m_Textures[_textureID];
 }
 
-void TextureManager::DrawTexture(std::string _textureID, int _xPos, int _yPos, int _width, int _height, SDL_RendererFlip _flip)
+void TextureManager::DrawTexture(std::string _textureID, float _xPos, float _yPos, int _width, int _height, SDL_RendererFlip _flip)
 {
-	SDL_Rect srcRect = { 0, 0, _width, _height };
-	SDL_Rect dstRect = { _xPos, _yPos, _width, _height };
+	SDL_Rect srcRect = { 0, 0, (float)_width, (float)_height };
+	SDL_Rect dstRect = { _xPos, _yPos, (float)_width, (float)_height };
 	SDL_RenderCopyEx(Game::GetInstance()->GetRenderer(), m_Textures[_textureID], &srcRect, &dstRect, 0, nullptr, _flip);
 }
 
@@ -64,14 +64,14 @@ void TextureManager::DrawTexture(std::string _textureID, int _xPos, int _yPos, i
 /// <param name="_frameColumn">column of frame to display</param>
 /// <param name="_flip">facing direction of sprite</param>
 /// <param name="scale">scale of frame</param>
-void TextureManager::DrawFrame(std::string _textureID, int _xPos, int _yPos, int _width, int _height, int _rows, int _cols, int _frameRow, int _frameColumn, SDL_RendererFlip _flip, float scale)
+void TextureManager::DrawFrame(std::string _textureID, float _xPos, float _yPos, int _width, int _height, int _rows, int _cols, int _frameRow, int _frameColumn, SDL_RendererFlip _flip, float scale)
 {
-	int frameWidth = _width / _cols;
-	int frameHeight = _height / _rows;
+	float frameWidth = (float)_width / _cols;
+	float frameHeight = (float)_height / _rows;
 
 	Vector2 frameStart = { frameWidth * (float)(_frameColumn - 1), frameHeight * (float)(_frameRow - 1) };
 
-	SDL_Rect srcRect = { (int)frameStart.x, (int)frameStart.y, frameWidth, frameHeight };
+	SDL_Rect srcRect = { frameStart.x, frameStart.y, frameWidth, frameHeight };
 	SDL_Rect dstRect = { _xPos, _yPos, frameWidth * scale, frameHeight * scale };
 	SDL_RenderCopyEx(Game::GetInstance()->GetRenderer(), m_Textures[_textureID], &srcRect, &dstRect, 0, nullptr, _flip);
 }
