+++ SDLGame2/SDLGame2/TextureManager.h
@@ -13,9 +13,9 @@ public:
 
 	SDL_Texture* LoadTexture(std::string _textureID);
 
-	void DrawTexture(std::string _textureID, int _xPos, int _yPos, int _width, int _height, SDL_RendererFlip _flip);
+	void DrawTexture(std::string _textureID, float _xPos, float _yPos, int _width, int _height, SDL_RendererFlip _flip);
 
-	void DrawFrame(std::string _textureID, int _xPos, int _yPos, int _width, int _height, int _rows, int _cols, int _frameRow, int frameColumn, SDL_RendererFlip _flip, float scale = 1);
+	void DrawFrame(std::string _textureID, float _xPos, float _yPos, int _width, int _height, int _rows, int _cols, int _frameRow, int frameColumn, SDL_RendererFlip _flip, float scale = 1);
 
 	//void SetRenderer(SDL_Renderer& _renderer);
